+++ components/selector/selectorMultiple/SelectorMultiple.tsx
@@ -1,11 +1,12 @@
 import { DropDownArrowIcon } from "@/icons/DropDownArrowIcon";
-import { Option, MultipleSelectorProps } from "../selector.interface";
+import { MultipleSelectorProps } from "../selector.interface";
 import { SCSelectorContainer, SCSelectorButton, SCSelectorOptions, SCSelectorOption, SCArrowButton, SCChipsWrapper } from "../selector.styles";
-import { useRef, useState, useEffect, FC } from "react";
+import { FC } from "react";
 import { SCWrapperSpinner } from "@/components/spinner/spinner.styles";
 import { SpinnerComponent } from "@/components/spinner/Spinner";
 import { Chip } from '@/components/chip/Chip';
 import { messages } from "@/messages/messages";
+import { useSelectMultiple } from "./useSelectorMultiple";
 
 
 
@@ -16,48 +17,7 @@ export const SelectorMultiple: FC<MultipleSelectorProps> = ({
   hasMore,
   label,
 }) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const [selectedValues, setSelectedValues] = useState<string[]>([]);
-  const containerRef = useRef<HTMLDivElement>(null);
-
-  const handleToggle = () => {
-    setIsOpen(!isOpen);
-  };
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (containerRef.current && !containerRef.current.contains(event.target as Node)) {
-      setIsOpen(false);
-    }
-  };
-
-  const handleSelect = (option: Option) => {
-    const optionValue = option.value as string;
-    const newSelectedValues = selectedValues.includes(optionValue)
-      ? selectedValues.filter(v => v !== optionValue)
-      : [...selectedValues, optionValue];
-      
-    setSelectedValues(newSelectedValues);
-    
-    if (onSelect) {
-      onSelect(newSelectedValues);
-    }
-  };
-  
-  const handleRemoveChip = (valueToRemove: string) => {
-    const newSelectedValues = selectedValues.filter(v => v !== valueToRemove);
-    setSelectedValues(newSelectedValues);
-    
-    if (onSelect) {
-      onSelect(newSelectedValues);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, []);
+const { isOpen, selectedValues, containerRef, handleToggle, handleSelect, handleRemoveChip } = useSelectMultiple({ onSelect });
 
   return (
     <SCSelectorContainer ref={containerRef}>
