+++ components/selector/selectorSimple/SelectorSimple.tsx
@@ -1,10 +1,10 @@
 import { DropDownArrowIcon } from "@/icons/DropDownArrowIcon";
-import { Option, SingleSelectorProps } from "../selector.interface";
+import { SingleSelectorProps } from "../selector.interface";
 import { SCSelectorContainer, SCSelectorButton, SCSelectorOptions, SCSelectorOption, SCArrowButton } from "../selector.styles";
-import { useRef, useState, useEffect } from "react";
 import { SpinnerComponent } from "@/components/spinner/Spinner";
 import { SCWrapperSpinner } from "@/components/spinner/spinner.styles";
 import { labels, messages } from "@/messages/messages";
+import { useSelectorSimple } from "./useSelectorSimple";
 
 
 
@@ -14,35 +14,8 @@ export const SelectorSimple = ({
   onSelect, 
   hasMore 
 }: SingleSelectorProps & { hasMore?: boolean }) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const [selected, setSelected] = useState<string | number>('');
-  const containerRef = useRef<HTMLDivElement>(null);
-
-  const handleToggle = () => {
-    setIsOpen(!isOpen);
-  };
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (containerRef.current && !containerRef.current.contains(event.target as Node)) {
-      setIsOpen(false);
-    }
-  };
-
-  const handleSelect = (option: Option) => {
-    setSelected(option.value as string);
-    handleToggle();
-    if (onSelect) {
-      onSelect(option.value as string);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, []);
-
+  
+const { isOpen, selected, containerRef, handleToggle, handleSelect } = useSelectorSimple({ onSelect });
   return (
     <SCSelectorContainer ref={containerRef}>
       <SCSelectorButton isOpen={isOpen} onClick={handleToggle}>
